feat: Update new transaction dialog
